backend/pyproject.toml
[tool.poetry]
name = "simplified-chinese-flashcards-backend"
version = "0.1.0"
description = "FastAPI backend for the Simplified Chinese Flashcards app"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
aiosqlite = "^0.19.0"
python-multipart = "^0.0.9"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
